+++ tests/testthat/helper.R
@@ -2,6 +2,7 @@ auth_sucess <- tryCatch(
   khisr:::khis_cred_testing(),
   khis_cred_internal_error = function(e) NULL
 )
+
 if(!isTRUE(auth_sucess)) {
   khisr:::khis_info(c(
     "!" = "Internal auth failed; calling {.fun khis_cred_clear}."
@@ -10,5 +11,5 @@ if(!isTRUE(auth_sucess)) {
 }
 
 skip_if_no_cred <- function() {
-  testthat::skip_if_not(khis_has_cred(), "No KHIS credentials")
+  testthat::skip_if_not(khis_has_cred(), "No DHIS2 credentials")
 }